agentic_security_system/requirements.txt
<<<<<<< HEAD
# Project Dependencies
# Agentic Security System

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=0.24.0
xgboost>=1.5.0
catboost>=1.0.0

# Configuration
pyyaml>=5.4.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Utilities
python-dotenv>=0.19.0
=======
openai
groq
langchain
langgraph
pydantic
numpy 
pandas
python-dotenv
langchain_groq
watchdog
psutil
scikit-learn
xgboost
matplotlib
seaborn
imbalanced-learn
joblib
numpy
>>>>>>> cb645a8f52bb31144c71a23275e733ff461db612
